Modified to exclude Highways owner from PROXY_OWNER exor user deletion
git-svn-id: svn://127.0.0.1/Core@12516 17c802ae-cbca-534c-9e8e-6d19a4926c45

+++ trunk/install/nm_4800_fix2.sql
@@ -1,11 +1,11 @@
 ----------------------------------------------------------------------------------------------------
 --   PVCS Identifiers :-
 --
---       PVCS id          : $Header:   //new_vm_latest/archives/nm3/install/nm_4800_fix2.sql-arc   1.1   Feb 18 2020 14:48:44   Chris.Baugh  $
+--       PVCS id          : $Header:   //new_vm_latest/archives/nm3/install/nm_4800_fix2.sql-arc   1.2   Feb 26 2020 11:50:34   Chris.Baugh  $
 --       Module Name      : $Workfile:   nm_4800_fix2.sql  $ 
---       Date into PVCS   : $Date:   Feb 18 2020 14:48:44  $
---       Date fetched Out : $Modtime:   Feb 18 2020 14:42:52  $
---       Version     	  : $Revision:   1.1  $
+--       Date into PVCS   : $Date:   Feb 26 2020 11:50:34  $
+--       Date fetched Out : $Modtime:   Feb 26 2020 11:49:16  $
+--       Version     	  : $Revision:   1.2  $
 --
 ----------------------------------------------------------------------------------------------------
 --   Copyright (c) 2020 Bentley Systems Incorporated. All rights reserved.
@@ -29,7 +29,7 @@ SET TERM ON
 --
 -- Spool to Logfile
 --
-DEFINE logfile1='nm_4800_fix1_&log_extension'
+DEFINE logfile1='nm_4800_fix2_&log_extension'
 spool &logfile1
 --
 --------------------------------------------------------------------------------
@@ -85,7 +85,7 @@ BEGIN
   --
   FOR l_rec IN c1 LOOP
     --
-    EXECUTE IMMEDIATE 'DELETE FROM hig_users WHERE hus_username = '''||l_rec.hus_username||'''';
+    EXECUTE IMMEDIATE 'DELETE FROM hig_users WHERE hus_username = '''||l_rec.hus_username||''' AND hus_is_hig_owner_flag != ''Y''';
     EXECUTE IMMEDIATE 'DELETE FROM hig_user_roles WHERE hur_username = '''||l_rec.hus_username||''' AND hur_role = ''PROXY_OWNER''';
     --
   END LOOP;
@@ -132,7 +132,7 @@ SET FEEDBACK OFF
 BEGIN
 	--
 	hig2.upgrade(p_product        => 'NET'
-				,p_upgrade_script => 'log_nm_4800_fix1.sql'
+				,p_upgrade_script => 'log_nm_4800_fix2.sql'
 				,p_remarks        => 'NET 4800 FIX 2 (Build 1)'
 				,p_to_version     => NULL
 				);
